requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM clients
anthropic==0.7.8
openai==1.3.7
httpx==0.25.2

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Utilities
python-multipart==0.0.6
jinja2==3.1.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
